+++ noodles-vcf/src/header/format.rs
@@ -39,12 +39,12 @@ impl Format {
     /// ```
     /// use noodles_vcf::header::{
     ///     format::{Key as FormatKey, Type},
-    ///     record::{Key as RecordKey, Value},
+    ///     record::{self, Value},
     ///     FileFormat, Format, Number, Record,
     /// };
     ///
     /// let record = Record::new(
-    ///     RecordKey::Format,
+    ///     record::key::FORMAT,
     ///     Value::Struct(vec![
     ///         (String::from("ID"), String::from("GT")),
     ///         (String::from("Number"), String::from("1")),
@@ -68,7 +68,7 @@ impl Format {
         file_format: FileFormat,
     ) -> Result<Self, TryFromRecordError> {
         match record.into() {
-            (record::Key::Format, record::Value::Struct(fields)) => {
+            (record::key::FORMAT, record::Value::Struct(fields)) => {
                 parse_struct(file_format, fields)
             }
             _ => Err(TryFromRecordError::InvalidRecord),
@@ -195,7 +195,7 @@ impl From<Key> for Format {
 impl fmt::Display for Format {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str(record::PREFIX)?;
-        f.write_str(record::Key::Format.as_ref())?;
+        f.write_str(record::key::FORMAT.as_ref())?;
         f.write_str("=<")?;
 
         write!(f, "{}={}", ID, self.id)?;
@@ -327,7 +327,7 @@ mod tests {
 
     fn build_record() -> Record {
         Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from("1")),
@@ -365,7 +365,7 @@ mod tests {
     #[test]
     fn test_try_from_record_file_format() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from(".")),
@@ -403,7 +403,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_extra_fields() -> Result<(), &'static str> {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from("1")),
@@ -437,7 +437,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_an_invalid_record_key() {
         let record = Record::new(
-            record::Key::FileFormat,
+            record::key::FILE_FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from("1")),
@@ -455,7 +455,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_an_invalid_record_value() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::String(String::from("VCFv4.3")),
         );
 
@@ -467,7 +467,7 @@ mod tests {
 
     #[test]
     fn test_try_from_record_for_format_with_a_missing_field() {
-        let record = Record::new(record::Key::Format, record::Value::Struct(Vec::new()));
+        let record = Record::new(record::key::FORMAT, record::Value::Struct(Vec::new()));
 
         assert_eq!(
             Format::try_from(record),
@@ -478,7 +478,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_an_invalid_id() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("")),
                 (String::from("Number"), String::from("1")),
@@ -496,7 +496,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_an_invalid_number() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from("NA")),
@@ -514,7 +514,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_an_invalid_type() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from("1")),
@@ -532,7 +532,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_an_invalid_idx() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from("1")),
@@ -551,7 +551,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_a_number_mismatch() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from(".")),
@@ -572,7 +572,7 @@ mod tests {
     #[test]
     fn test_try_from_record_for_format_with_a_type_mismatch() {
         let record = Record::new(
-            record::Key::Format,
+            record::key::FORMAT,
             record::Value::Struct(vec![
                 (String::from("ID"), String::from("GT")),
                 (String::from("Number"), String::from("1")),